requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]==2.8.2
httpx==0.27.2
pillow==10.4.0
reportlab==4.2.2
openpyxl==3.1.5
pyotp==2.9.0
aiofiles==24.1.0
google-cloud-storage==2.18.2
google-cloud-firestore==2.17.0
qrcode[pil]==8.0
cryptography==43.0.1
bcrypt==4.2.0
